edit periodical transfers enum fix.
git-svn-id: http://dev.ezhecom.com/svn/pbc@54 179d6c2c-9243-4f44-b038-fef1748950ba

+++ web/Private/EditPeriodicalTransfer.aspx.cs
@@ -252,7 +252,7 @@ namespace AIM.PBC.Web.Private.Pages
 			if(entity.PeriodType == PeriodTypes.Standard)
 			{
 				SelectStandardPeriod();
-				rblStandardPeriod.Items.FindByValue(entity.StandardPeriod.ToString()).Selected = true;
+				rblStandardPeriod.Items.FindByValue(((byte)entity.StandardPeriod).ToString()).Selected = true;
 			}
 			else
 			{
